+++ src/localization/natto_wheel_odometry/src/swerve_odometry.cpp
@@ -40,7 +40,7 @@ swerve_odometry::swerve_odometry (const rclcpp::NodeOptions &node_options) : Nod
     RCLCPP_INFO (this->get_logger (), "swerve_odometry node has been initialized.");
     RCLCPP_INFO (this->get_logger (), "wheel_radius: %.2f m", wheel_radius_);
     RCLCPP_INFO (this->get_logger (), "Number of wheels: %d", num_wheels_);
-    for (int i = 0; i < num_wheels_; i++) {
+    for (size_t i = 0; i < num_wheels_; i++) {
         RCLCPP_INFO (this->get_logger (), "wheel_position_xy[%d]: (%.2f, %.2f)", i, wheel_position_x_[i], wheel_position_y_[i]);
     }
     RCLCPP_INFO (this->get_logger (), "publish_tf : %s", publish_tf_ ? "true" : "false");
@@ -60,7 +60,7 @@ void swerve_odometry::swerve_callback (const natto_msgs::msg::Swerve::SharedPtr 
     double ATA[3][3] = {};  // A^T * A
     double ATb[3]    = {};  // A^T * b
 
-    for (int i = 0; i < num_wheels_; i++) {
+    for (size_t i = 0; i < num_wheels_; i++) {
         double angle = msg->wheel_angle[i];
         double speed = msg->wheel_speed[i] * 2.0 * M_PI * wheel_radius_;
